feat: unstable version
Took 21 hours 7 minutes

+++ src/main/java/com/pawscodes/sierras/kafka/bitrix/data/entity/StockData.java
@@ -26,5 +26,5 @@ public class StockData {
     int ano;
     @Id
     int mes;
-    long stock;
+    double stock;
 }
